Split robot load into its own plan
