+++ src/mx_bluesky/hyperion/device_setup_plans/utils.py
@@ -20,6 +20,7 @@ def fill_in_energy_if_not_supplied(dcm: DCM, detector_params: DetectorParams):
     if not detector_params.expected_energy_ev:
         actual_energy_ev = 1000 * (yield from bps.rd(dcm.energy_in_kev))
         detector_params.expected_energy_ev = actual_energy_ev
+    return detector_params
 
 
 def start_preparing_data_collection_then_do_plan(
