+++ tests/test_data/parameter_json_files/good_test_robot_load_and_centre_params.json
@@ -1,0 +1,21 @@
+{
+    "parameter_model_version": "5.0.0",
+    "zocalo_environment": "dev_artemis",
+    "beamline": "BL03S",
+    "insertion_prefix": "SR03S",
+    "storage_directory": "/tmp/",
+    "visit": "cm31105-4",
+    "file_name": "file_name",
+    "run_number": 0,
+    "use_roi_mode": false,
+    "det_dist_to_beam_converter_path": "tests/test_data/test_lookup_table.txt",
+    "omega_start_deg": 0,
+    "transmission_frac": 1.0,
+    "exposure_time_s": 0.004,
+    "detector_distance_mm": 255,
+    "demand_energy_ev": 11100,
+    "sample_id": 12345,
+    "sample_puck": 40,
+    "sample_pin": 3,
+    "comment": "Descriptive comment."
+}
